+++ cmake/tests/unit/gluecodium_generate/cpp-internal-namespace/cpp/FooImpl.cpp
@@ -21,9 +21,9 @@
 #include "unit/test/Foo.h"
 
 namespace unit::test {
-std::shared_ptr<Foo>
-Foo::make_foo(const ::custom::common::optional< double >& parameter)
+::custom::common::Return< std::shared_ptr<Foo>, std::error_code >
+Foo::make_foo(const std::optional< double >& parameter)
 {
-    return nullptr;
+    return {std::shared_ptr<Foo>{}};
 }
 } // namespace unit::test