+++ gluecodium/src/test/resources/smoke/external_types/output/cbridge/src/smoke/cbridge_ExternalInterface.cpp
@@ -7,11 +7,11 @@
 #include "cbridge_internal/include/TypeInitRepository.h"
 #include "cbridge_internal/include/WrapperCache.h"
 #include "foo/Bar.h"
-#include "gluecodium/Optional.h"
 #include "gluecodium/TypeRepository.h"
 #include <cstdint>
 #include <memory>
 #include <new>
+#include <optional>
 #include <string>
 void smoke_ExternalInterface_release_handle(_baseRef handle) {
     delete get_pointer<::std::shared_ptr< ::smoke::ExternalInterface >>(handle);
@@ -98,17 +98,17 @@ smoke_ExternalInterface_SomeStruct_release_handle( _baseRef handle )
 _baseRef
 smoke_ExternalInterface_SomeStruct_create_optional_handle(_baseRef someField)
 {
-    auto _struct = new ( ::std::nothrow ) ::gluecodium::optional<::smoke::ExternalInterface::some_Struct>( ::smoke::ExternalInterface::some_Struct( ) );
+    auto _struct = new ( ::std::nothrow ) std::optional<::smoke::ExternalInterface::some_Struct>( ::smoke::ExternalInterface::some_Struct( ) );
     (*_struct)->some_Field = Conversion<::std::string>::toCpp( someField );
     return reinterpret_cast<_baseRef>( _struct );
 }
 _baseRef
 smoke_ExternalInterface_SomeStruct_unwrap_optional_handle( _baseRef handle )
 {
-    return reinterpret_cast<_baseRef>( &**reinterpret_cast<::gluecodium::optional<::smoke::ExternalInterface::some_Struct>*>( handle ) );
+    return reinterpret_cast<_baseRef>( &**reinterpret_cast<std::optional<::smoke::ExternalInterface::some_Struct>*>( handle ) );
 }
 void smoke_ExternalInterface_SomeStruct_release_optional_handle(_baseRef handle) {
-    delete reinterpret_cast<::gluecodium::optional<::smoke::ExternalInterface::some_Struct>*>( handle );
+    delete reinterpret_cast<std::optional<::smoke::ExternalInterface::some_Struct>*>( handle );
 }
 _baseRef smoke_ExternalInterface_SomeStruct_someField_get(_baseRef handle) {
     auto struct_pointer = get_pointer<const ::smoke::ExternalInterface::some_Struct>(handle);