+++ gluecodium/src/test/resources/smoke/lambdas/output/cpp/include/smoke/Lambdas.h
@@ -4,11 +4,11 @@
 // -------------------------------------------------------------------------------------------------
 #pragma once
 #include "gluecodium/ExportGluecodiumCpp.h"
-#include "gluecodium/Optional.h"
 #include "gluecodium/UnorderedMapHash.h"
 #include "gluecodium/VectorHash.h"
 #include <cstdint>
 #include <functional>
+#include <optional>
 #include <string>
 #include <unordered_map>
 #include <vector>
@@ -25,7 +25,7 @@ public:
     using Confuser = ::std::function<::smoke::Lambdas::Producer(const ::std::string&)>;
     using Consumer = ::std::function<void(const ::std::string&)>;
     using Indexer = ::std::function<int32_t(const ::std::string&, const float)>;
-    using NullableConfuser = ::std::function<::gluecodium::optional< ::smoke::Lambdas::Producer >(const ::gluecodium::optional< ::std::string >&)>;
+    using NullableConfuser = ::std::function<std::optional< ::smoke::Lambdas::Producer >(const std::optional< ::std::string >&)>;
 public:
     virtual ::smoke::Lambdas::Producer deconfuse( const ::std::string& value, const ::smoke::Lambdas::Confuser& confuser ) = 0;
     static ::std::unordered_map< int32_t, ::std::string > fuse( const ::std::vector< ::std::string >& items, const ::smoke::Lambdas::Indexer& callback );