+++ gluecodium/src/test/resources/smoke/listeners/output/dart/ffi/ffi_smoke_CalculatorListener.cpp
@@ -195,7 +195,7 @@ FfiOpaqueHandle
 library_smoke_CalculatorListener_ResultStruct_create_handle_nullable(FfiOpaqueHandle value)
 {
     return reinterpret_cast<FfiOpaqueHandle>(
-        new (std::nothrow) gluecodium::optional<smoke::CalculatorListener::ResultStruct>(
+        new (std::nothrow) std::optional<smoke::CalculatorListener::ResultStruct>(
             gluecodium::ffi::Conversion<smoke::CalculatorListener::ResultStruct>::toCpp(value)
         )
     );
@@ -203,13 +203,13 @@ library_smoke_CalculatorListener_ResultStruct_create_handle_nullable(FfiOpaqueHa
 void
 library_smoke_CalculatorListener_ResultStruct_release_handle_nullable(FfiOpaqueHandle handle)
 {
-    delete reinterpret_cast<gluecodium::optional<smoke::CalculatorListener::ResultStruct>*>(handle);
+    delete reinterpret_cast<std::optional<smoke::CalculatorListener::ResultStruct>*>(handle);
 }
 FfiOpaqueHandle
 library_smoke_CalculatorListener_ResultStruct_get_value_nullable(FfiOpaqueHandle handle)
 {
     return gluecodium::ffi::Conversion<smoke::CalculatorListener::ResultStruct>::toFfi(
-        **reinterpret_cast<gluecodium::optional<smoke::CalculatorListener::ResultStruct>*>(handle)
+        **reinterpret_cast<std::optional<smoke::CalculatorListener::ResultStruct>*>(handle)
     );
 }
 FfiOpaqueHandle